requirements.txt
absl-py==0.13.0
asgiref==3.4.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
Django==3.2.4
dm-tree==0.1.6
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gym==0.18.3
h5py==3.1.0
idna==2.10
imageio==2.9.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.2.5
Pillow==8.2.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.15
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.7.0
six==1.15.0
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-probability==0.13.0
termcolor==1.1.0
tifffile==2021.6.14
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
